+++ libyafuse2/libext4/super.c
@@ -101,13 +101,13 @@ int32_t ext4_fill_super_info(struct super_block *sb, struct ext4_super_block *es
 
 int32_t ext4_raw_super(struct ext4_super_block *sb)
 {
-  int32_t offset = 0;
-  size_t sb_sz = 0;
+  int64_t offset = 0;
+  int64_t sb_sz = 0;
   int32_t ret = 0;
 
   offset = EXT4_GROUP_0_PAD_SZ;
 
-  ret = io_seek((long)offset);
+  ret = io_seek(offset);
   if (ret != 0) {
     return -1;
   }
@@ -117,7 +117,7 @@ int32_t ext4_raw_super(struct ext4_super_block *sb)
    * default size of superblock is 1024 bytes
    */
   sb_sz = sizeof(struct ext4_super_block);
-  ret = io_read((uint8_t *)sb, (size_t)sb_sz);
+  ret = io_read((uint8_t *)sb, sb_sz);
   if (ret != 0) {
     memset((void *)sb, 0, sb_sz);
     return -1;
